enhancement (search&action modules): move the JdnDatepicker directive to @knora/action and update public_api + NgModule
