+++ projects/knora/action/src/public_api.ts
@@ -10,6 +10,7 @@ export * from './lib/admin-image/admin-image.directive';
 export * from './lib/directives/gnd.directive';
 export * from './lib/directives/mathjax.directive';
 export * from './lib/existing-name/existing-name.directive';
+export * from './lib/directives/jdn-datepicker.directive';
 
 export * from './lib/action.module';
